requirements.txt
# ============================================
# CORE DEPENDENCIES
# ============================================

# Data processing
pandas==2.0.3
numpy==1.24.3

# Web scraping
playwright==1.39.0
beautifulsoup4==4.12.2
requests==2.31.0

# Database clients
psycopg2-binary==2.9.7
sqlalchemy==1.4.47
duckdb==0.10.0
pyarrow==13.0.0

# ============================================
# AIRFLOW & ORCHESTRATION
# ============================================

# Airflow is pre-installed in base image, only install extra providers
apache-airflow-providers-postgres==5.7.1
pendulum>=2.1.2,<3.0

# ============================================
# STORAGE & INTEGRATION
# ============================================

# MinIO integration
minio==7.2.3

# ============================================
# MONITORING & APIs
# ============================================

# Monitoring (optional - Phase 2)
# prometheus-client==0.19.0
# psutil==5.9.7

# Dashboard (optional - Phase 2)
# streamlit==1.29.0

# ============================================
# CODE QUALITY & TESTING
# ============================================

pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0

# ============================================
# UTILITIES
# ============================================

python-dotenv==1.0.0
PyYAML==6.0.1